Editing down, replacing outline with text
removed some large swaths of outline.  Also stripped out some over-long text describing manorialism (as much as I can remember it).  I figure we'll defer to the tour section for a lot of that.

Wrote out a description of the Commonwealth Council (and removed that section of outline).  More description of the Council Agencies would be nice.

Removed the notable places, since those are summarized elsewhere.