Editing
Cleaned up the results of the python script that converted the aside blocks.
Caught one that was a div instead of an aside for some reason.

Hit a few of Sovereign's edit suggestions, including some editing for confidence and clarity, and catching a few emberassing examples of loose/lose and wrote/rote.