Monastery test 2 (#24)
* Update 35-module-monastery.md

Tweaks and additions to module

* Added a heading stub for the Secret Army

* Added in a dialog snippet CV done for the Drakes

* Added an initial draft for poppy juice

* resolved merge conflicts, added in a ton of content

Resolved some merge conflicts.  I'm sure git will explode later.

Added a bunch of content for the monastery module.

* Editing

Cleaned up the results of the python script that converted the aside blocks.
Caught one that was a div instead of an aside for some reason.

Hit a few of Sovereign's edit suggestions, including some editing for confidence and clarity, and catching a few emberassing examples of loose/lose and wrote/rote.

* just called them cost categories

Jumped back from editing character creation to see what I called the item categories, saw I just called them "categories," thought that was too vague, so I edited it to just call them "cost categories."

I've probably called them "cost categories" everywhere else already anyway.

* updated clippings with some more snipped text

updated clippings with some text removed from Character Creation.

Because I never throw code away, apparently.

* reformatting/rewriting chargen

Chargen has been an ugly bullet list for way too long; now it's not.

Also fixed some auto-generated aside blocks.

Kicked around the wording of the advancement part for way too long, ultimately used Morgan's version.

* the character to have used extensively

better...?

* more clippings

* Editing, formatting, rearranging

Fixed some custom tags that had internal headers.

Moved some elements out of the summary and into the body for the Kingdoms of Men

* Turned the priest block into text

Turned the priest bit into a section (it was a bullet list)

Removed a bit at the beginning that should be... somewhere else, I'm not sure where.  I think I put it in the clip-file.

* de-bullet-listed

Turned this chapter into paragraphs of text (it was still a big-ass bullet list)

* edit, fixed a custom block

Fixed a custom-block that had an internal header.

Removed a word in an edit.

* minor edit

changed "motivation" to "pick your motivations" in a steps-list.

* chewing

I've continued editing and tinkering with this section.

* minor editing

* shifted chapters around

Moved Chapter 18: representing a character to  Chapter 11, so that it comes after the basic rules.

Pulled "powers" out of chapter 15, added a chapter 16 for "features"

* designing and using powers

The "designing and using powers" text has been damned to the clipfile.

It needs to go somewhere, just not here, and I think it's redundant to text taht exists in the powers chapter anyway.

* editing and shifting-about

Much editing and shifting-about.

* a chapter on features

Moving the rules for features into their own chapter

* removed a design note

It's going to the Features section.

* fixed an auto-genned custom tag

fixed a custom block with an internal heading

* fixed some auto-genne'd custom tags

* split magic and luck into two chapters

* initial content for chapter

* renamed to just magic

* minor editing

* minor edits in Shaded Sea

* minor edits, work on The Bridge encounter

* Moved a comment into the clippings file.

* De-bullet-listed the ranged attacks section.

* Added soe basic check rules to the quick-reference.  (At Sovereign's request.)

* Worked on the River encounter in the Monastery.

* Filled in some missing sections.

* Filled in some on the Leapers, some minor edits in the Guardians

* Added script to split chapters

* rewrite the Special Cases block, removed Pole Arms and Disable Device, minor edits

After feedback, removed Pole Arms and Disable Device.
Rewrote the block at the end; still working on it.
Changed a few lines to match the formatting of the rest of the source.

* add en-dash

* Expanded the note, so taht the moons are named, and we explicitly mention that there are planets.

* Expanded the Terrace-Builders a bit

* Copied in the text of CV's Stormsword's Rest write up.  Need to actually edit and integrate it.

* Initial import of the Secret Army / Wolf-in-Wool's Flock.

Oops, I said "added the Secret Army content" when I committed the Organizations chapter.
I actually just slightly edited the Far Watch.

* Updated the change list to start over; added credit for CV for the Secret Army and Stormsword's Rest.

* Still getting the Stormsword's Rest content in here

* In the Kingdoms, replaced a bunch of tripple-dashes with unicode en-dashes, added line-feeds between sentences (like the rest of the document), and tweaked some wording about the Kobold genocide.

* Made a hole, added an initial import of Stormsword's Rest (thanks CV!)

* minor edit to Kingdoms tagline.